Refactored updateSwaggerSecurityDefinitionForStore
