+++ components/apimgt/org.wso2.carbon.apimgt.impl/src/main/java/org/wso2/carbon/apimgt/impl/definitions/AsyncApiParser.java
@@ -28,13 +28,13 @@ import org.osgi.service.component.annotations.Component;
 import org.wso2.carbon.apimgt.api.APIDefinition;
 import org.wso2.carbon.apimgt.api.APIDefinitionValidationResponse;
 import org.wso2.carbon.apimgt.api.APIManagementException;
+import org.wso2.carbon.apimgt.api.dto.KeyManagerConfigurationDTO;
 import org.wso2.carbon.apimgt.api.model.API;
 import org.wso2.carbon.apimgt.api.model.APIProduct;
 import org.wso2.carbon.apimgt.api.model.Scope;
 import org.wso2.carbon.apimgt.api.model.SwaggerData;
 import org.wso2.carbon.apimgt.api.model.URITemplate;
-import org.wso2.carbon.apimgt.impl.APIConstants;
-import org.wso2.carbon.apimgt.impl.utils.APIUtil;
+import org.wso2.carbon.apimgt.impl.definitions.APIConstants;
 
 import java.net.URI;
 import java.net.URISyntaxException;
@@ -1630,7 +1630,7 @@ public class AsyncApiParser extends APIDefinition {
         if (!opScopes.isEmpty()) {
             if (opScopes.size() == 1) {
                 String firstScope = opScopes.get(0);
-                Scope scope = APIUtil.findScopeByKey(scopes, firstScope);
+                Scope scope = OASParserUtil.findScopeByKey(scopes, firstScope);
                 if (scope == null) {
                     throw new APIManagementException("Scope '" + firstScope + "' not found.");
                 }
@@ -1638,7 +1638,7 @@ public class AsyncApiParser extends APIDefinition {
                 template.setScopes(scope);
             } else {
                 for (String scopeName : opScopes) {
-                    Scope scope = APIUtil.findScopeByKey(scopes, scopeName);
+                    Scope scope = OASParserUtil.findScopeByKey(scopes, scopeName);
                     if (scope == null) {
                         throw new APIManagementException("Resource Scope '" + scopeName + "' not found.");
                     }
@@ -1875,14 +1875,17 @@ public class AsyncApiParser extends APIDefinition {
     }
 
     @Override
-    public String getOASDefinitionForStore(API api, String oasDefinition,
-            Map<String, String> hostsWithSchemes, String kmId) throws APIManagementException {
+    public String getOASDefinitionForStore(API api, String oasDefinition, Map<String, String> hostsWithSchemes,
+                                           KeyManagerConfigurationDTO keyManagerConfigurationDTO)
+            throws APIManagementException {
         return null;
     }
 
     @Override
     public String getOASDefinitionForStore(APIProduct product, String oasDefinition,
-            Map<String, String> hostsWithSchemes, String kmId) throws APIManagementException {
+                                           Map<String, String> hostsWithSchemes,
+                                           KeyManagerConfigurationDTO keyManagerConfigurationDTO)
+            throws APIManagementException {
         return null;
     }
