+++ components/apimgt/org.wso2.carbon.apimgt.impl/src/main/java/org/wso2/carbon/apimgt/impl/definitions/OAS3Parser.java
@@ -58,7 +58,6 @@ import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.json.simple.JSONObject;
-import org.wso2.carbon.apimgt.api.APIAdmin;
 import org.wso2.carbon.apimgt.api.APIDefinition;
 import org.wso2.carbon.apimgt.api.APIDefinitionValidationResponse;
 import org.wso2.carbon.apimgt.api.APIManagementException;
@@ -72,10 +71,8 @@ import org.wso2.carbon.apimgt.api.model.CORSConfiguration;
 import org.wso2.carbon.apimgt.api.model.Scope;
 import org.wso2.carbon.apimgt.api.model.SwaggerData;
 import org.wso2.carbon.apimgt.api.model.URITemplate;
-import org.wso2.carbon.apimgt.impl.APIAdminImpl;
-import org.wso2.carbon.apimgt.impl.APIConstants;
+import org.wso2.carbon.apimgt.impl.definitions.APIConstants;
 import org.wso2.carbon.apimgt.impl.utils.APIUtil;
-import org.wso2.carbon.context.PrivilegedCarbonContext;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -92,7 +89,7 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import java.util.UUID;
 
-import static org.wso2.carbon.apimgt.impl.APIConstants.APPLICATION_JSON_MEDIA_TYPE;
+import static org.wso2.carbon.apimgt.impl.definitions.APIConstants.APPLICATION_JSON_MEDIA_TYPE;
 import static org.wso2.carbon.apimgt.impl.definitions.OASParserUtil.isValidWithPathsWithTrailingSlashes;
 
 /**
@@ -929,17 +926,19 @@ public class OAS3Parser extends APIDefinition {
      * @param api            API
      * @param oasDefinition  OAS definition
      * @param hostsWithSchemes host addresses with protocol mapping
-     * @param kmId             UUID of the Key Manager
+     * @param keyManagerConfigurationDTO configuration details of the Key Manager
      * @return OAS definition
      */
     @Override
-    public String getOASDefinitionForStore(API api, String oasDefinition,
-            Map<String, String> hostsWithSchemes, String kmId) throws APIManagementException {
+    public String getOASDefinitionForStore(API api, String oasDefinition, Map<String, String> hostsWithSchemes,
+                                           KeyManagerConfigurationDTO keyManagerConfigurationDTO)
+            throws APIManagementException {
 
         OpenAPI openAPI = getOpenAPI(oasDefinition);
         updateOperations(openAPI);
         updateEndpoints(api, hostsWithSchemes, openAPI);
-        return updateSwaggerSecurityDefinitionForStore(openAPI, new SwaggerData(api), hostsWithSchemes, kmId);
+        return updateSwaggerSecurityDefinitionForStore(openAPI, new SwaggerData(api), hostsWithSchemes,
+                keyManagerConfigurationDTO);
     }
 
     /**
@@ -948,17 +947,20 @@ public class OAS3Parser extends APIDefinition {
      * @param product        APIProduct
      * @param oasDefinition  OAS definition
      * @param hostsWithSchemes host addresses with protocol mapping
-     * @param kmId             UUID of the Key Manager
+     * @param keyManagerConfigurationDTO configuration details of the Key Manager
      * @return OAS definition
      */
     @Override
     public String getOASDefinitionForStore(APIProduct product, String oasDefinition,
-            Map<String, String> hostsWithSchemes, String kmId) throws APIManagementException {
+                                           Map<String, String> hostsWithSchemes,
+                                           KeyManagerConfigurationDTO keyManagerConfigurationDTO)
+            throws APIManagementException {
 
         OpenAPI openAPI = getOpenAPI(oasDefinition);
         updateOperations(openAPI);
         updateEndpoints(product, hostsWithSchemes, openAPI);
-        return updateSwaggerSecurityDefinitionForStore(openAPI, new SwaggerData(product), hostsWithSchemes, kmId);
+        return updateSwaggerSecurityDefinitionForStore(openAPI, new SwaggerData(product), hostsWithSchemes,
+                keyManagerConfigurationDTO);
     }
 
     /**
@@ -1470,30 +1472,14 @@ public class OAS3Parser extends APIDefinition {
      * @param openAPI        OpenAPI
      * @param swaggerData    SwaggerData
      * @param hostsWithSchemes GW hosts with protocols
-     * @param kmId             UUID of the Key Manager
+     * @param keyManagerConfigurationDTO configuration details of the Key Manager
      * @return updated OAS definition
      * @throws APIManagementException if an error occurred
      */
     private String updateSwaggerSecurityDefinitionForStore(OpenAPI openAPI, SwaggerData swaggerData,
-            Map<String,String> hostsWithSchemes, String kmId) throws APIManagementException {
-
-        KeyManagerConfigurationDTO keyManagerConfigurationDTO = null;
-        try {
-            if (!StringUtils.isEmpty(kmId)) {
-                String tenantDomain = PrivilegedCarbonContext.getThreadLocalCarbonContext().getTenantDomain();
-                APIAdmin apiAdmin = new APIAdminImpl();
-                keyManagerConfigurationDTO = apiAdmin.getKeyManagerConfigurationById(tenantDomain, kmId);
-                if (keyManagerConfigurationDTO == null || (StringUtils.isEmpty(kmId) && !Objects.equals(
-                        keyManagerConfigurationDTO.getType(), APIConstants.KeyManager.DEFAULT_KEY_MANAGER_TYPE))) {
-                    keyManagerConfigurationDTO = apiAdmin.getKeyManagerConfigurationByName(tenantDomain,
-                            APIConstants.KeyManager.DEFAULT_KEY_MANAGER);
-                }
-            }
-        } catch (APIManagementException e) {
-            throw new APIManagementException("Failed to retrieve key manager information by name or ID",
-                    ExceptionCodes.ERROR_RETRIEVE_KM_INFORMATION);
-        }
-
+                                                           Map<String, String> hostsWithSchemes,
+                                                           KeyManagerConfigurationDTO keyManagerConfigurationDTO)
+            throws APIManagementException {
         String authUrl;
         // By Default, add the GW host with HTTPS protocol if present.
         if (hostsWithSchemes.containsKey(APIConstants.HTTPS_PROTOCOL)) {
